SimpleShift Python API/requirements.txt
flask
sqlalchemy
pymysql
cryptography
sender
flask_cors
passlib
Flask-JWT
werkzeug
flask_jwt_extended
datetime
python-dotenv
